status code

+++ src/routes/todoRoutes.js
@@ -1,12 +1,13 @@
 const conn = require('../config/database')
 const router = require('express').Router()
+const auth = require('../config/auth')
 
 // Create todo
-router.post('/user/todo', (req, res) => {
-   // req.body = {user_id, description}
+router.post('/todo', auth, (req, res) => {
+   // req.body = {description}
 
    const sql = `INSERT INTO todos SET ?`
-   const data = req.body
+   const data = { user_id : req.user.id , description : req.body.description}
 
    conn.query(sql, data, (err, result) => {
       // Jika terjadi masalah ketika running sql nya.
@@ -19,11 +20,11 @@ router.post('/user/todo', (req, res) => {
    })
 })
 
-// Exercise
-// Read todo
-router.get('/user/todo/:userid', (req, res) => {
+
+// Read all todo
+router.get('/todo', auth, (req, res) => {
    const sql = `SELECT * FROM todos WHERE user_id = ?`
-   const data = req.params.userid
+   const data = req.user.id
 
    conn.query(sql, data, (err, result) => {
       if(err) return res.send(err.sqlMessage)
@@ -33,13 +34,13 @@ router.get('/user/todo/:userid', (req, res) => {
 })
 
 // Update todo
-router.patch('/user/todo/:todoid', (req, res) => {
-   const sql = `UPDATE todos SET ? WHERE id = ?`
+router.patch('/todo/:todoid', auth, (req, res) => {
+   const sql = `UPDATE todos SET ? WHERE id = ? AND user_id = ?`
    // Jika menggunakan tanda tanya (escape query) lebih dari satu, variable 'data' harus berupa array
    // Dimana array tersebut berisi data yang akan me-replace tanda tanya yang ada. Urutan itu diperhitungkan
-   const data = [req.body, req.params.todoid]
+   const data = [req.body, req.params.todoid, req.user.id]
 
-   conn.query(sql, data, (err, result) => {
+   conn.query(sql, data, (error, result) => {
       if(err) return res.send(err.sqlMessage)
 
       res.send({
@@ -49,9 +50,9 @@ router.patch('/user/todo/:todoid', (req, res) => {
 })
 
 // Delete todo
-router.delete('/user/todo/:todoid', (req, res) => {
-   const sql = `DELETE FROM todos WHERE id = ?`
-   const data = req.params.todoid
+router.delete('/todo/:todoid', auth, (req, res) => {
+   const sql = `DELETE FROM todos WHERE id = ? AND user_id = ?`
+   const data = [req.params.todoid, req.user.id]
 
    conn.query(sql, data, (err, result) => {
       if(err) return res.send(err.sqlMessage)
